refactor: remove lots of code from iroh-bytes::get
this is all not really related to the iroh-bytes protocol but only relevant
for iroh itself. So in order to keep the surface area small, let's move it
into iroh.
