+++ iroh/Cargo.toml
@@ -15,6 +15,7 @@ rust-version = "1.65"
 [dependencies]
 anyhow = { version = "1", features = ["backtrace"] }
 bao-tree = { version = "0.5.0-rc.1", features = ["tokio_fsm"], default-features = false }
+blake3 = "1.3.3"
 derive_more = { version = "0.99.17", git = "https://github.com/JelteF/derive_more", features = ["debug", "display", "from", "try_into"] }
 flume = "0.10.14"
 futures = "0.3.25"
